requirements.txt
fastapi
python-dotenv
mysql.connector